Allow for 'parse' to accept raw input (e.g. string or Buffer) (#60)
* Adding test cases for checking raw input

Two cases were added ('helloworld-as-buffer.spec.js' and
'helloworld-as-string.spec.js') were added as a way to validate that
the 'helloworld.spec.js' test would pass if the test file's contents
were loaded before passing it into the 'parse' function.

* Transferred RAML parse check to a function

The check in '_sourceToRamlObj' to see if the parsed object is RAML 1.0
and can be expanded will be re-used if there are multiple kinds of
inputs that can be parsed (e.g. the content in a string). As such, this
has been exported to another function for future re-use.

* Added checks and parsing for raw RAML content

New parsing functionality has been implemented by assuming the source
is raw content after attempting to find if it exists or if it is a
Buffer. To keep the validation of the generated RAML object consistent,
it uses the '_validateLoadedRaml' function.